+++ packages/amplify-cli/src/__tests__/extensions/amplify-helpers/get-category-pluginInfo.test.ts
@@ -3,7 +3,7 @@ import { constructMockPluginPlatform } from './mock-plugin-platform';
 import { constructContext } from '../../../context-manager';
 import { getCategoryPluginInfo } from '../../../extensions/amplify-helpers/get-category-pluginInfo';
 
-test('getCategoryPluginInfo', () => {
+test('getCategoryPluginInfo returns the first pluginInfo to match category', () => {
   const mockPluginPlatform = constructMockPluginPlatform();
   const mockProcessArgv = [
     '/Users/userName/.nvm/versions/node/v12.16.1/bin/node',
@@ -15,5 +15,35 @@ test('getCategoryPluginInfo', () => {
   const mockInput = new Input(mockProcessArgv);
   const mockContext = constructContext(mockPluginPlatform, mockInput);
   const hostingPluginInfo = getCategoryPluginInfo(mockContext, 'hosting');
+  expect(hostingPluginInfo).toBeDefined();
+});
+
+test('getCategoryPluginInfo returns pluginInfo when plugin matches category and service', () => {
+  const mockPluginPlatform = constructMockPluginPlatform();
+  const mockProcessArgv = [
+    '/Users/userName/.nvm/versions/node/v12.16.1/bin/node',
+    '/Users/userName/.nvm/versions/node/v12.16.1/bin/amplify',
+    'hosting',
+    'add',
+  ];
+
+  const mockInput = new Input(mockProcessArgv);
+  const mockContext = constructContext(mockPluginPlatform, mockInput);
+  const hostingAmplifyhostingPluginInfo = getCategoryPluginInfo(mockContext, 'hosting', 'amplifyhosting');
+  expect(hostingAmplifyhostingPluginInfo).toBeDefined();
+});
+
+test('getCategoryPluginInfo returns the first pluginInfo to match only category when no service match exists', () => {
+  const mockPluginPlatform = constructMockPluginPlatform();
+  const mockProcessArgv = [
+    '/Users/userName/.nvm/versions/node/v12.16.1/bin/node',
+    '/Users/userName/.nvm/versions/node/v12.16.1/bin/amplify',
+    'hosting',
+    'add',
+  ];
+
+  const mockInput = new Input(mockProcessArgv);
+  const mockContext = constructContext(mockPluginPlatform, mockInput);
+  const hostingPluginInfo = getCategoryPluginInfo(mockContext, 'hosting', 'S3');
   expect(hostingPluginInfo).toBeDefined();
 });