+++ packages/amplify-category-hosting/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.7.4](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-hosting@2.7.3...amplify-category-hosting@2.7.4) (2021-04-27)
+
+**Note:** Version bump only for package amplify-category-hosting
+
+
+
+
+
 ## [2.7.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-hosting@2.7.2...amplify-category-hosting@2.7.3) (2021-04-19)
 
 **Note:** Version bump only for package amplify-category-hosting